python_Text_extraction/requirements.txt
certifi==2021.10.8
cycler==0.11.0
editdistpy==0.1.3
fonttools==4.28.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
opencv-python==4.5.5.62
pdf2image==1.16.0
Pillow==9.0.0
pytesseract==0.3.8
python-dateutil==2.8.2
scipy==1.7.3
symspellpy==6.7.6
tqdm==4.62.3







bs4
requests
urllib3
selenium
tqdm
textract
xmi-reader
dkpro-cassis

certifi==2021.10.8
pdf2image==1.16.0
Pillow==8.4.0
pytesseract==0.3.8

